Make d3.component() interface conform to selection API
The default selection API passes in [data, index, parentIndex].

We first chose [data, props], which prevents us from passing in indices. An alternative would be [props, data, index, parentIndex], but that would change the API significantly.

In the end, selection.props() was chosen to access props. It's similar to selection.data(), also storing the props on the DOM node in __props__.