Add tooltips and improvements to the pie charts.
Adding tooltips to the pie charts includes a refactoring of the pie chart examples
to pull more setup logic out of the render function and into the time either before
data is loaded or immediately after data is loaded. Also adds sszvis.format.percent
for formatting percentages, and this commit adds a little white dot in the center of
pie charts, to match the design.