Create custom layout algorithms for the vertical and the horizontal bar charts.
These algorithms are encapsulated inside two helper functions on sszvis.fn.
This commit also includes reimplementations of all bar chart examples which use the new layouts,
in addition to readjusting of chart sizes to ensure optimal display.

NOTE: the layout algorithm may not be stable, nor is the API for the two helper functions necessarily
the final version. This is why these have not been fully documented in the source.