Change the way base layer maps are rendered to be able to add the lake to some of them
Incidentally, this removes the grundkarte renderer component, which wasn't flexible enough. Instead, the base layers are now composed. Also adds the ability to custom specify the border color of the border above the patterned lake overlay. This is used by the base layer components to ensure that the border above the lake matches the color of the base layer borders