In the axis class, create oft-used selections once, and reuse them
This improves the quality of the codebase by reducing oft-used expressions to single variables, it increases performance by avoiding re-creating the same selection many times, and it fixes a bug where the axis title was being moved a bit as if it was an axis label, which was caused by an inexact selection of all 'text' elements. The single construction of these selections relies on the invariant that new elements are not added later in the render script.