Change the way the map components handle doing the data merge
This enables adding anchored shapes over the maps (for example, in the signature bubble chart). This involves moving the data merge step outside of the base map component, which removes the keyName property from the base map, and adds a mergedData property. the keyName property is now used in the map modules themselves, and they pass pre-merged data into the various components which they compose together. Also adds the optional anchoredShapes property, which allows you to specify a shape anchor component which will be rendered over the other components.