Add link labels to the sankey chart
The array of links which should have source labels, and the array of links which should have target labels, should be passed into the component. This is done as an optimization. Rather than saving an id somewhere, and then doing a lookup in the component's data, it makes much more sense to just pass in this information from outside, since it's already there as part of the node hover lookup. If necessary, it's possible to save an id and filter the link data for labeled links outside the component.