Fix some little issues in the linear color scale rendering
This fixes an issue where the ending circle was a different color than the final bar. Indeed, the final bar didn’t capture the full range of the color scale.

It also fixes an issue where the circles at either end of the color scale weren’t being correctly re-rendered and updated.