Add back in part of the previous geoJson optimization
The part that was added back in is to cache the centroid of all the geojson objects. But whereas before the centroid was cached in screen space, now it’s cached in spherical coordinates. It’s still reprojected each time the map is rendered, but this saves the browser from needing to do the same centroid calculations many times.