Bar chart move behavior revisions
Adds functionality to the move behavior to support bar chart “profile panning” on mobile, where the user sees a tooltip and the move behavior cancels scrolling only when the touch event happens within the space on the chart occupied by the bars, and not elsewhere. This creates a virtual “profile” shape that matches the heights of the bars, within which touch events will show a tooltip and allow the user to move around, but outside of which the user scrolls as usual.

Implemented in the horizontal “interactive” example and in the grouped bar chart “two-small” example, the latter of which displays some rather tricky acrobatics required to correctly calculate the “profile” of groujps of bars which can have positive or negative values. Includes also a utility function for helping with the “profile” test. This utility function is used in the horizontal “interactive” example.