Revisions to certain bar charts’ interactions
Not all interactive bar charts were revised, and not all of these revisions were performed in the same way. In each case, I made a choice between three options. First, the move behavior, configured to compute a correct “profile” for the bar chart which blocks scrolling on that profile. The best example of where this is suitable is on the horizontally-oriented bar charts. Second, the move behavior without any special profile configuration, which allows for left-right panning but allows the browser to scroll on up-down touch movement. This is suitable for dense vertical bar charts, where the user just wants to pan left-right to see all the tooltips. Third, the panning behavior, which unlike the move behavior doesn’t show tooltips in the spaces between bars, but which is suitable for situations where vertical movement is expected, but the profile approach would cover too much of the screen in a scroll trap. The best example of this is the vertical and horizontal stacked bar charts. The “profile” for these charts covers the entire chart area, so the only good place to enable scrolling is in the gaps between the bars.