feat: Prevent color legend circles from being cut-off
We shift the circle by a pixel to prevent cut-off. We could've also
changed the translation of every legend, but decided to do this for the
whole library – the worst that will happen is that in charts where this
had been fixed through translation, the legend will be 1px off. We
accept this as a trade-off to fix this problem in charts that don't deal
with this explicitely.